düzenlemeler yapıldı.

+++ IssueProject/Models/User/LoginInfo.cs
@@ -6,7 +6,7 @@ namespace IssueProject.Models.User
     {
         public int UserId { get; set; }
         public string FullName { get; set; }
-        public short Role { get; set; }
+        public int Role { get; set; }
 
         public string AccessToken { get; set; } = "";
         public string RefreshToken { get; set; } = "";
